Split vscode out of diagnostics parsing utils.
Diagnostics parsing is just parsing the error line number and column number from the error string
There is no need depends on vscode to do that.
So remove the depends on vscode from these error message parsing functions,
So these functions can be tested without vscode(by nodejs directly).
After this, the tests can be running in a faster way and easier debugging.
