+++ src/diagnostics/msvc.ts
@@ -1,10 +1,9 @@
 /**
  * Module for handling MSVC diagnostics
- */ /** */
+ */
 
-import * as vscode from 'vscode';
-
-import { oneLess, RawDiagnosticParser, FeedLineResult } from './util';
+import { oneLess } from '@cmt/basic/util';
+import { FeedLineResult, RawDiagnosticParser } from './rawDiagnosticParser';
 
 export const REGEX = /^\s*(\d+>)?\s*([^\s>].*)\((\d+|\d+,\d+|\d+,\d+,\d+,\d+)\)\s*:\s+((?:fatal )?error|warning|info)\s*(\w{1,2}\d+)?\s*:\s*(.*)$/;
 
@@ -19,17 +18,17 @@ export class Parser extends RawDiagnosticParser {
             const parts = location.split(',');
             const n0 = oneLess(parts[0]);
             if (parts.length === 1) {
-                return new vscode.Range(n0, 0, n0, 999);
+                return {startLine: n0, startCharacter: 0, endLine: n0, endCharacter: 999};
             }
             if (parts.length === 2) {
                 const n1 = oneLess(parts[1]);
-                return new vscode.Range(n0, n1, n0, n1);
+                return {startLine: n0, startCharacter: n1, endLine: n0, endCharacter: n1};
             }
             if (parts.length === 4) {
                 const n1 = oneLess(parts[1]);
                 const n2 = oneLess(parts[2]);
                 const n3 = oneLess(parts[3]);
-                return new vscode.Range(n0, n1, n2, n3);
+                return {startLine: n0, startCharacter: n1, endLine: n2, endCharacter: n3};
             }
             throw new Error('Unable to determine location of MSVC diagnostic');
         })();
