+++ src/drivers/driver.ts
@@ -13,7 +13,7 @@ import { ArgsCompileCommand } from '@cmt/compdb';
 import { ConfigurationReader, defaultNumJobs } from '@cmt/config';
 import { CMakeBuildConsumer, CompileOutputConsumer } from '@cmt/diagnostics/build';
 import { CMakeOutputConsumer } from '@cmt/diagnostics/cmake';
-import { RawDiagnosticParser } from '@cmt/diagnostics/util';
+import { RawDiagnosticParser } from '@cmt/diagnostics/rawDiagnosticParser';
 import { ProgressMessage } from '@cmt/drivers/cms-client';
 import * as expand from '@cmt/expand';
 import { CMakeGenerator, effectiveKitEnvironment, Kit, kitChangeNeedsClean, KitDetect, getKitDetect, getVSKitEnvironment } from '@cmt/kit';
