+++ src/util.ts
@@ -10,6 +10,8 @@ import { DebuggerEnvironmentVariable, execute } from '@cmt/proc';
 import rollbar from '@cmt/rollbar';
 import { Environment, EnvironmentUtils } from './environmentVariables';
 
+export * from "@cmt/basic/util";
+
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
 
@@ -171,70 +173,6 @@ export function isTruthy(value: (boolean | string | null | undefined | number)) 
     }
     // Numbers/bools/etc. follow common C-style truthiness
     return !!value;
-}
-
-/**
- * Generate an array of key-value pairs from an object using
- * `getOwnPropertyNames`
- * @param obj The object to iterate
- */
-export function objectPairs<V>(obj: { [key: string]: V }): [string, V][] {
-    return Object.getOwnPropertyNames(obj).map(key => ([key, obj[key]] as [string, V]));
-}
-
-/**
- * Map an iterable by some projection function
- * @param iter An iterable to map
- * @param proj The projection function
- */
-export function* map<In, Out>(iter: Iterable<In>, proj: (arg: In) => Out): Iterable<Out> {
-    for (const item of iter) {
-        yield proj(item);
-    }
-}
-
-export function* chain<T>(...iter: Iterable<T>[]): Iterable<T> {
-    for (const sub of iter) {
-        for (const item of sub) {
-            yield item;
-        }
-    }
-}
-
-export function reduce<In, Out>(iter: Iterable<In>, init: Out, mapper: (acc: Out, el: In) => Out): Out {
-    for (const item of iter) {
-        init = mapper(init, item);
-    }
-    return init;
-}
-
-export function find<T>(iter: Iterable<T>, predicate: (value: T) => boolean): T | undefined {
-    for (const value of iter) {
-        if (predicate(value)) {
-            return value;
-        }
-    }
-    // Nothing found
-    return undefined;
-}
-
-/**
- * Generate a random integral value.
- * @param min Minimum value
- * @param max Maximum value
- */
-export function randint(min: number, max: number): number {
-    return Math.floor(Math.random() * (max - min) + min);
-}
-
-export function product<T>(arrays: T[][]): T[][] {
-    return arrays.reduce(
-        (acc, curr) =>
-            // Append each element of the current array to each list already accumulated
-            acc.map(prev => curr.map(item => prev.concat(item)))
-                // Join all the lists
-                .reduce((a, b) => a.concat(b), []),
-        [[]] as T[][]);
 }
 
 export interface CMakeValue {
